# api/openapi.yaml
openapi: 3.1.0
info:
  title: birdnet-go-remote-mic management API
  version: 0.1.0
  summary: Management API for the BirdNET-Go remote microphone appliance.
  description: |
    Read and manage a birdnet-go-remote-mic appliance: appliance and
    per-device status, and a narrow slice of configuration.

    The API is served over HTTPS with a self-signed certificate generated
    and persisted by the appliance. Consumers are BirdNET-Go (generated Go
    client) and the appliance's own web UI (generated TypeScript types).

    Errors use RFC 9457 problem details (application/problem+json).

    Authentication: when the appliance has a shared access token configured
    (auth.token), every operation except /healthz requires it as an RFC
    6750 bearer credential (Authorization: Bearer <token>). A missing or
    wrong token yields 401 with a WWW-Authenticate: Bearer challenge and an
    RFC 9457 problem body (covered by each operation's default response).
    Without a configured token the appliance serves every operation
    unauthenticated (open access, the default), and the mDNS TXT record
    says auth=none rather than auth=token. The same token is the Digest
    password for the RTSP stream.

    Evolution policy: additive changes only within v1. Fields may be added
    to responses at any time; consumers must ignore unknown fields.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
  contact:
    name: Tomi P. Hakala
...
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /status:
    get:
      operationId: getStatus
      tags: [system]
      summary: Appliance status
      description: >
        Appliance-level runtime state: version, uptime, RTSP listener,
        discovery, whether an access token is required, and per-device counts.
        Full per-device detail lives at /devices.
      responses:
        "200":
          description: Current appliance status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplianceStatus"
        default:
          $ref: "#/components/responses/Problem"
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List capture devices
      description: >
        Every configured device with its configured and negotiated
        parameters and runtime state, including devices that failed to open
        (state "skipped") or died after startup (state "failed").
      responses:
        "200":
          description: All configured devices.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: provisionDevice
      tags: [devices]
      summary: Enable (provision) a detected capture device
      description: >
        Enables streaming on a capture device the host exposes but the
        configuration does not yet list. The appliance derives a unique name
        from the device's hardware label, generates a hard-to-guess RTSP path,
        and chooses a mode, sample rate and channel count from the device's
        probed capabilities (Opus 48 kHz mono when supported, otherwise raw PCM
        at the device's best rate). Any field set in the request overrides its
        derived default. The change is persisted and hot-applied to the running
        pipeline; the new device starts serving without a restart.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionDeviceRequest"
      responses:
        "201":
          description: The device was provisioned and is being served.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: No such capture device is present on the host.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: >
            The device is already configured: by this exact id or derived name,
            or under another id (the host lists it but the available view hides
            it, for example a card index that resolves to it).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: The request is well-formed but fails validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblem"
        default:
          $ref: "#/components/responses/Problem"
...
  /devices/{name}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get one capture device
      description: The named device with its configuration and runtime state.
      parameters:
        - name: name
          in: path
          required: true
          description: The device's configured unique name.
          schema:
            type: string
      responses:
        "200":
          description: The named device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Remove a configured capture device
      description: >
        Removes the named device from the configuration, stopping it if it is
        serving and returning its hardware to the pool of available devices. The
        change is persisted and hot-applied without a restart.
      parameters:
        - name: name
          in: path
          required: true
          description: The device's configured unique name.
          schema:
            type: string
      responses:
        "204":
          description: The device was removed.
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /events:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Live event stream
      description: |
        A single multiplexed Server-Sent Events stream for the whole
        appliance. Each SSE event's `event:` field names its type and its
        `data:` field is a JSON document:

        - `levels`: emitted at a fixed cadence (10 Hz) with one
          LevelsEvent covering every serving device. Peak and RMS are
          measured over the window since the previous event; meter
          ballistics (decay, peak hold) are the client's job.
        - `notification`: emitted when a notification is raised, with one
          Notification as data. The stream is best effort: there is no replay,
          `Last-Event-ID` is ignored, and a client re-syncs from GET
          /notifications on connect, on reconnect, and whenever it detects an
          id gap. An active condition's onset whose title or message changed
          while it stayed active is sent again under its own id; a client
          replaces the entry it holds with that id (it is not a new entry).
        - `heartbeat`: emitted every 15 seconds with `{}` as data, so
          idle connections stay alive and clients can detect a dead
          server. Sent as a named event (not an SSE comment) so simple
          fetch-based parsers see it.

        Further event types will be added additively; clients must ignore
        event types they do not know. Level data is ephemeral: there is no
        replay and `Last-Event-ID` is ignored.

        Browser note: the native EventSource API cannot send an
        Authorization header, so an authenticated web client should
        consume this endpoint with fetch and a streamed response body.

        OpenAPI 3.1 cannot type per-event payloads on a stream; the
        `data:` schema for `levels` events is LevelsEvent. This operation
        moves to OpenAPI 3.2 `itemSchema` once the code generators
        support it.
      parameters:
        - name: events
          in: query
          required: false
          description: >
            Event types to subscribe to, comma-separated. Absent means
            all types. Heartbeats are always sent.
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
              examples: [levels]
      responses:
        "200":
          description: The event stream; stays open until the client disconnects.
          content:
            text/event-stream:
              schema:
                type: string
                description: >
                  SSE framing per the HTML living standard. The data field
                  of each `levels` event is a JSON LevelsEvent.
        default:
          $ref: "#/components/responses/Problem"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      description: >
        The current notification snapshot: the boot identity, the server's
        wall-clock time and uptime read at one instant, the ring depth, the
        next id that will be assigned, and every discrete
        history entry still in the ring merged with every active condition (even
        ones the ring has trimmed), in ascending id order. Clients bootstrap and
        re-sync from this; the /events stream is best effort, so a gap in
        streamed ids means a dropped event and the client refetches here. The
        ring keeps the most recent 500 history entries in memory only: there is
        no time-based expiry, and a restart (a new bootId) starts it empty. A
        full snapshot runs to about 165 KB of JSON, so the response is
        gzip-compressed when the request's Accept-Encoding allows it.
      responses:
        "200":
          description: The current notification snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSnapshot"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        The appliance's shared access token (auth.token in the
        configuration). Enforced on every operation except /healthz once a
        token is configured; the mDNS TXT record advertises auth=token or
        auth=none accordingly. The same token is the Digest password for
        the RTSP stream.
  responses:
    Problem:
      description: An RFC 9457 problem detail.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ApplianceStatus:
      type: object
      description: Appliance-level runtime state.
      required:
        - version
        - uptimeSeconds
        - rtspListen
        - discoveryEnabled
        - authRequired
        - devicesServing
        - devicesTotal
      properties:
        version:
          type: string
          description: Appliance build version (set at build time).
        uptimeSeconds:
          type: integer
          format: int64
          minimum: 0
          description: Seconds since the process started.
        rtspListen:
          type: string
          description: The RTSP listener address (host:port as configured).
          examples: [":8554"]
        discoveryEnabled:
          type: boolean
          description: Whether mDNS/DNS-SD advertisement is on.
        authRequired:
          type: boolean
          description: >
            Whether a shared access token is configured, so the API, web
            UI and RTSP stream require credentials.
        devicesServing:
          type: integer
          minimum: 0
          description: Devices currently capturing and serving their track.
        devicesTotal:
          type: integer
          minimum: 0
          description: Devices in the configuration, whatever their state.
        overrides:
          type: array
          description: >
            Serve-time CLI flag overrides in force for this run. Each entry
            names a config field whose effective (running) value differs from
            the persisted config file value because a serve flag (--listen,
            --mgmt-listen, --cert-dir, --management, --discovery) overrode it.
            Absent or empty when no serve overrides are active.
          items:
            $ref: "#/components/schemas/ConfigOverride"
    ConfigOverride:
      type: object
      description: >
        One config field overridden by a serve CLI flag for this run. The
        effective value is in force now; the persisted value is what the config
        file holds and what a restart without the flag would use.
      required: [field, effective, persisted]
      properties:
        field:
          type: string
          description: Dotted config path of the overridden field.
          examples: ["listen", "management.listen"]
        effective:
          type: string
          description: Value in force for this run (from the serve flag).
        persisted:
          type: string
          description: Value stored in the config file.
...
    ProvisionDeviceRequest:
      type: object
      description: >
        Request to enable (provision) a detected capture device. Only device is
        required; the appliance derives sensible defaults for everything else,
        and any field set here overrides its derived default.
      required:
        - device
      properties:
        device:
          type: string
          maxLength: 2048
          description: Capture device id to enable, as reported by GET /devices/available.
          examples: ["usb:1235:8218:s=S1A2B3:if=0,0"]
        name:
          type: string
          maxLength: 128
          description: Optional unique device name; derived from the hardware label when omitted.
        mode:
          $ref: "#/components/schemas/StreamMode"
        rate:
          type: integer
          minimum: 8000
          maximum: 384000
          description: Optional capture sample rate in Hz; chosen from the device's capabilities when omitted.
        channels:
          type: array
          maxItems: 8
          items:
            type: integer
            minimum: 1
            maximum: 8
          description: >
            Optional 1-based channel selection to stream, ascending and unique.
            When omitted the appliance selects a single channel: it captures
            briefly from the device and picks the loudest channel (the lowest
            numbered among channels within 1 dB of the loudest, and channel 1
            when all are near silence or the measurement fails). An empty array
            is treated the same as omitting the field (so no minItems is
            imposed).
    StreamMode:
      type: string
      enum: [pcm, opus]
      description: >
        pcm streams raw L16 at the capture rate (the ultrasonic path);
        opus streams 48 kHz Opus, mono or stereo (one or two channels; the
        normal-audio path).
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
    ValidationProblem:
      description: A problem detail carrying per-field validation errors.
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, reason]
                properties:
                  field:
                    type: string
                    description: Dotted path of the invalid field.
                    examples: ["devices[0].rate"]
                  reason:
                    type: string
...